+++ src/helpers/element-ref.js
@@ -1,17 +1,4 @@
-export function validate(ref, warning) {
-    const isValid =
-        ref instanceof Element ||
-        (ref && ref._isVue) ||
-        typeof ref === 'string';
-
-    if (!isValid && warning) {
-        console.warn(warning);
-    }
-
-    return isValid;
-}
-
-export function resolve(ref, fallback) {
+export function resolveRef(ref, fallback) {
     if (ref instanceof Element) {
         return ref;
     } else if (ref && ref._isVue) {
@@ -22,8 +9,3 @@ export function resolve(ref, fallback) {
 
     return fallback;
 }
-
-export default {
-    validate,
-    resolve,
-};